Refactored ContractorWeeklyProgress
